connect_four::game
Module player

This module contains the Player struct, which represents a player in the game.

Each player has a unique name and token. The token is a single character that represents the player on the game board.

Example

use connect_four::Player;

let player = Player::new("Alice");
println!("Player name: {}", player.name);
println!("Player token: {}", player.token);

Structs

Player

Represents a player in the game.

Statics

NEXT_TOKEN 🔒

The next token to be assigned to a player.